requirements.txt
psycopg2-binary~=2.9.10
pytest~=8.3.2
mockito~=1.5.0